requirements.txt
RPi.GPIO
paho-mqtt
influxdb
